pyproject.toml
[tool.poetry]
name = "coding_challenge"
version = "0.0.0"
description = "Coding Challenge"
authors = ["Dominik Kozaczko <dominik@kozaczko.info>"]
license = "GPL-3.0"
readme = "README.md"
homepage = "https://github.com/dekoza/coding_challenge"
repository = "https://github.com/dekoza/coding_challenge"
documentation = "https://coding_challenge.readthedocs.io"
classifiers = [
    "Development Status :: 3 - Alpha",
]

[tool.poetry.urls]
Changelog = "https://github.com/dekoza/coding_challenge/releases"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
hypercorn = {version = "^0.14.3", extras = ["trio", "uvloop"]}
anyio = "^3.6.2"

#[tool.poetry.scripts]
#coding_challenge = "coding_challenge.__main__:main"
typer = "^0.6.1"
tortoise-orm = {version = "^0.19.2", extras = ["accel", "asyncpg"]}
httpx = "^0.23.0"
python-dotenv = "^0.21.0"
spotipy = "^2.20.0"
pendulum = "^2.1.2"

[tool.poetry.group.dev.dependencies]
Pygments = ">=2.10.0"
black = "^22.10.0"
coverage = {extras = ["toml"], version = ">=6.2"}
darglint = ">=1.8.1"
furo = ">=2021.11.12"
isort = "^5.10.1"
mypy = "^0.982"
myst-parser = {version = ">=0.16.1"}
pep8-naming = ">=0.12.1"
pre-commit = "^2.20.0"
pre-commit-hooks = "^4.3.0"
pytest = ">=6.2.5"
pyupgrade = ">=2.29.1"
ruff = "^0.0.83"
safety = ">=1.10.3"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
ipython = "^8.5.0"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["coding_challenge", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.isort]
profile = "black"

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[tool.ruff]
line-length = 88
select = [
    "F401",
    "F403",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
